Text Annotation Frames (#8055)
