+++ common/api/summary/core-backend.exports.csv
@@ -4,6 +4,9 @@ public;interface;AcquireNewBriefcaseIdArg
 public;class;AnnotationElement2d
 preview;class;AnnotationElement2d
 beta;type;AnyDb
+beta;function;appendFrameToBuilder
+beta;function;appendTextAnnotationGeometry
+beta;interface;AppendTextAnnotationGeometryArgs
 public;class;AuxCoordSystem
 preview;class;AuxCoordSystem
 public;class;AuxCoordSystem2d
@@ -68,10 +71,16 @@ internal;interface;FontIndexProps
 public;class;CodeSpecs
 preview;class;CodeSpecs
 public;interface;ComputedProjectExtents
+beta;function;computeFrame
+beta;interface;ComputeFrameArgs
 beta;function;computeGraphemeOffsets
 beta;interface;ComputeGraphemeOffsetsArgs
+beta;function;computeIntervalPoints
+beta;interface;ComputeIntervalPointsArgs
 beta;function;computeLayoutTextBlockResult
 public;interface;ComputeProjectExtentsOptions
+internal;type;ComputeRangesForTextLayout
+internal;interface;ComputeRangesForTextLayoutArgs
 alpha;type;ConcreteEntity
 alpha;type;ConcreteEntityProps
 beta;function;convertEC2SchemasToEC3Schemas
@@ -231,6 +240,8 @@ beta;class;ExternalSourceIsInRepository
 beta;class;ExternalSourceOwnsAttachments
 public;class;FileNameResolver
 deprecated;class;FileNameResolver
+internal;type;FindFontId
+internal;type;FindTextStyle
 beta;class;FolderContainsRepositories
 alpha;class;FolderLink
 beta;interface;FontFile
@@ -339,8 +350,10 @@ public;class;IpcHost
 public;interface;IpcHostOpts
 public;interface;ITwinIdArg
 public;class;KnownLocations
+beta;function;layoutTextBlock
 beta;interface;LayoutTextBlockArgs
 internal;class;LightLocation
+beta;class;LineLayout
 public;class;LineStyle
 preview;class;LineStyle
 public;namespace;LineStyleDefinition
@@ -411,8 +424,7 @@ public;class;PhysicalTypeIsOfPhysicalMaterial
 public;class;PlanCallout
 public;class;Platform
 internal;interface;ProcessChangesetOptions
-beta;function;produceTextAnnotationGeometry
-beta;interface;ProduceTextAnnotationGeometryArgs
+beta;function;produceTextBlockGeometry
 public;type;ProgressFunction
 public;enum;ProgressStatus
 public;type;PropertyHandler
@@ -444,6 +456,7 @@ preview;class;RoleElement
 public;class;RoleModel
 preview;class;RoleModel
 public;class;RpcTrace
+beta;class;RunLayout
 public;class;Schema
 public;interface;SchemaImportOptions
 internal;type;SchemaKey
@@ -549,6 +562,8 @@ public;class;TextAnnotation2d
 preview;class;TextAnnotation2d
 public;class;TextAnnotation3d
 preview;class;TextAnnotation3d
+beta;class;TextBlockLayout
+internal;interface;TextLayoutRanges
 public;class;Texture
 preview;class;Texture
 beta;interface;TextureCreateProps
