+++ common/api/summary/core-common.exports.csv
@@ -899,6 +899,8 @@ public;interface;TextAnnotation3dProps
 preview;interface;TextAnnotation3dProps
 beta;interface;TextAnnotationAnchor
 beta;interface;TextAnnotationCreateArgs
+beta;type;TextAnnotationFillColor
+beta;type;TextAnnotationFrameShape
 beta;interface;TextAnnotationProps
 beta;class;TextBlock
 beta;class;TextBlockComponent
@@ -910,6 +912,7 @@ beta;interface;TextBlockLayoutResult
 beta;interface;TextBlockMargins
 beta;interface;TextBlockProps
 beta;interface;TextBlockStringifyOptions
+beta;interface;TextFrameStyleProps
 beta;class;TextRun
 beta;interface;TextRunProps
 public;class;TextString
