+++ common/changes/@itwin/core-backend/claudiaareneee-annotation-frames_2025-05-28-15-47.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/core-backend",
+      "comment": "Refactoring the TextAnnotationGeometry and adding support for frames",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/core-backend"
+}
