+++ core/backend/src/annotations/TextBlockLayout.ts
@@ -8,9 +8,10 @@
 
 import { BaselineShift, FontId, FontType, FractionRun, LineLayoutResult, Paragraph, Run, RunLayoutResult, TextBlock, TextBlockLayoutResult, TextBlockMargins, TextRun, TextStyleSettings, TextStyleSettingsProps } from "@itwin/core-common";
 import { Geometry, Range2d } from "@itwin/core-geometry";
-import { IModelDb } from "./IModelDb";
+import { IModelDb } from "../IModelDb";
 import { assert, NonFunctionPropertiesOf } from "@itwin/core-bentley";
 import * as LineBreaker from "linebreak";
+
 
 /** @internal */
 export interface TextLayoutRanges {
@@ -69,7 +70,7 @@ export interface LayoutTextBlockArgs {
  * If the document specifies a width > 0, individual lines are split to try to avoid exceeding that width.
  * Individual TextRuns can be split onto multiple lines at word boundaries if necessary. Individual FractionRuns are never split.
  * @see [[computeLayoutTextBlockResult]]
- * @internal
+ * @beta
  */
 export function layoutTextBlock(args: LayoutTextBlockArgs): TextBlockLayout {
   const findFontId = args.findFontId ?? ((name, type) => args.iModel.fonts.findId({ name, type }) ?? 0);
@@ -84,7 +85,7 @@ export function layoutTextBlock(args: LayoutTextBlockArgs): TextBlockLayout {
 /**
  * Gets the result of laying out the the contents of a TextBlock into a series of lines containing runs.
  * The visual layout accounts for the [TextStyle]($common)s, fonts, and [TextBlock.width]($common). It applies word-wrapping if needed.
- * The layout returned matches the visual layout of the geometry produced by [[produceTextAnnotationGeometry]].
+ * The layout returned matches the visual layout of the geometry produced by [[appendTextAnnotationGeometry]].
  * @beta
  */
 export function computeLayoutTextBlockResult(args: LayoutTextBlockArgs): TextBlockLayoutResult {
@@ -290,7 +291,12 @@ function split(source: string): Segment[] {
   return segments;
 }
 
-/** @internal */
+/**
+ * Represents the layout of a single run (text, fraction, or line break) within a line of text.
+ * Stores information about the run's position, style, and font within the line.
+ * Provides utilities for splitting text runs for word wrapping and converting to result objects.
+ * @beta
+ */
 export class RunLayout {
   public source: Run;
   public charOffset: number;
@@ -422,7 +428,12 @@ export class RunLayout {
   }
 }
 
-/** @internal */
+/**
+ * Represents the layout of a single line within a paragraph of a text block.
+ * Contains a sequence of RunLayout objects, the computed range of the line, and its offset from the document origin.
+ * Provides utilities for appending runs, computing ranges, and converting to result objects.
+ * @beta
+ */
 export class LineLayout {
   public source: Paragraph;
   public range = new Range2d(0, 0, 0, 0);
@@ -492,7 +503,9 @@ export class LineLayout {
 
 /**
  * Describes the layout of a text block as a collection of lines containing runs.
- * @internal
+ * Computes the visual layout of the text block, including word wrapping, justification, and margins.
+ * Provides access to the computed lines, ranges, and utilities for converting to result objects.
+ * @beta
  */
 export class TextBlockLayout {
   public source: TextBlock;
@@ -522,7 +535,7 @@ export class TextBlockLayout {
   public toResult(): TextBlockLayoutResult {
     return {
       lines: this.lines.map((x) => x.toResult(this.source)),
-      range: this.textRange.toJSON(),
+      range: this.range.toJSON(),
     };
   }
