+++ core/common/src/annotation/TextAnnotation.ts
@@ -8,6 +8,7 @@
 
 import { Point3d, Range2d, Transform, XYZProps, YawPitchRollAngles, YawPitchRollProps } from "@itwin/core-geometry";
 import { TextBlock, TextBlockProps } from "./TextBlock";
+import { TextStyleColor } from "./TextStyle";
 
 /** Describes how to compute the "anchor point" for a [[TextAnnotation]].
  * The anchor point is a point on or inside of the 2d bounding box enclosing the contents of the annotation's [[TextBlock]].
@@ -35,6 +36,35 @@ export interface TextAnnotationAnchor {
   horizontal: "left" | "center" | "right";
 }
 
+/** Set of predefined shapes that can be computed and drawn around the margins of a [[TextBlock]]
+ * @beta
+*/
+export type TextAnnotationFrameShape = "none" | "line" | "rectangle" | "circle" | "equilateralTriangle" | "diamond" | "square" | "pentagon" | "hexagon" | "octagon" | "capsule" | "roundedRectangle";
+
+
+/**
+ * Describes what color to use when filling the frame around a [[TextBlock]].
+ * If `background` is specified, [[GeometryParams.BackgroundFill]] will be set to `BackgroundFill.Outline`.
+ * @beta
+ */
+export type TextAnnotationFillColor = TextStyleColor | "background";
+
+/**
+ * Describes how to draw the frame around a [[TextBlock]].
+ * The frame can be a simple line, a filled shape, or both.
+ * @beta
+ */
+export interface TextFrameStyleProps {
+  /** Shape of the frame. Default: "rectangle" */
+  shape?: TextAnnotationFrameShape;
+  /** The color to fill the shape of the text frame. This fill will is applied using [[FillDisplay.Blanking]]. Default: no fill */
+  fill?: TextAnnotationFillColor;
+  /** The color of the text frame's outline. Default: black */
+  border?: TextStyleColor;
+  /** This will be used to set the [[GeometryParams.weight]] property of the frame (in pixels). Default: 1px */
+  borderWeight?: number;
+}
+
 /**
  * JSON representation of a [[TextAnnotation]].
  * @beta
@@ -48,6 +78,8 @@ export interface TextAnnotationProps {
   textBlock?: TextBlockProps;
   /** See [[TextAnnotation.anchor]]. Default: top-left. */
   anchor?: TextAnnotationAnchor;
+  /** See [[TextAnnotation.frame]]. Default: no frame */
+  frame?: TextFrameStyleProps
 }
 
 /** Arguments supplied to [[TextAnnotation.create]].
@@ -62,6 +94,8 @@ export interface TextAnnotationCreateArgs {
   textBlock?: TextBlock;
   /** See [[TextAnnotation.anchor]]. Default: top-left. */
   anchor?: TextAnnotationAnchor;
+  /** See [[TextAnnotation.frame]]. Default: no frame */
+  frame?: TextFrameStyleProps
 }
 
 /**
@@ -74,7 +108,7 @@ export interface TextAnnotationCreateArgs {
  * edges, or in the center of the box.
  * - The [[orientation]] is applied to rotate the box around the anchor point.
  * - Finally, the [[offset]] is added to the anchor point to apply translation.
- * @see [produceTextAnnotationGeometry]($backend) to decompose the annotation into a set of geometric primitives suitable for use with [[GeometryStreamBuilder.appendTextBlock]].
+ * @see [appendTextAnnotationGeometry]($backend) to construct the geometry and append it to an [[ElementGeometry.Builder]].
  * @beta
  */
 export class TextAnnotation {
@@ -88,12 +122,15 @@ export class TextAnnotation {
   public anchor: TextAnnotationAnchor;
   /** An offset applied to the anchor point that can be used to position annotations within the same geometry stream relative to one another. */
   public offset: Point3d;
+  /** The frame settings of the text annotation. */
+  public frame?: TextFrameStyleProps;
 
-  private constructor(offset: Point3d, angles: YawPitchRollAngles, textBlock: TextBlock, anchor: TextAnnotationAnchor) {
+  private constructor(offset: Point3d, angles: YawPitchRollAngles, textBlock: TextBlock, anchor: TextAnnotationAnchor, frame?: TextFrameStyleProps) {
     this.offset = offset;
     this.orientation = angles;
     this.textBlock = textBlock;
     this.anchor = anchor;
+    this.frame = frame
   }
 
   /** Creates a new TextAnnotation. */
@@ -102,8 +139,11 @@ export class TextAnnotation {
     const angles = args?.orientation ?? new YawPitchRollAngles();
     const textBlock = args?.textBlock ?? TextBlock.createEmpty();
     const anchor = args?.anchor ?? { vertical: "top", horizontal: "left" };
+    // If the user supplies a frame, but doesn't supply a shape, default the shape to "rectangle"
+    const shape: TextAnnotationFrameShape = args?.frame?.shape ?? "rectangle";
+    const frame = args?.frame ? { shape, ...args.frame } : undefined;
 
-    return new TextAnnotation(offset, angles, textBlock, anchor);
+    return new TextAnnotation(offset, angles, textBlock, anchor, frame);
   }
 
   /**
@@ -115,6 +155,7 @@ export class TextAnnotation {
       orientation: props?.orientation ? YawPitchRollAngles.fromJSON(props.orientation) : undefined,
       textBlock: props?.textBlock ? TextBlock.create(props.textBlock) : undefined,
       anchor: props?.anchor ? { ...props.anchor } : undefined,
+      frame: props?.frame ? { shape: "rectangle", ...props.frame } : undefined,
     });
   }
 
@@ -139,6 +180,9 @@ export class TextAnnotation {
     if (this.anchor.vertical !== "top" || this.anchor.horizontal !== "left") {
       props.anchor = { ...this.anchor };
     }
+
+    // Default frame to "none"
+    props.frame = this.frame ? { ...this.frame } : undefined;
 
     return props;
   }
@@ -193,8 +237,14 @@ export class TextAnnotation {
 
   /** Returns true if this annotation is logically equivalent to `other`. */
   public equals(other: TextAnnotation): boolean {
+    const framesMatch = this.frame?.shape === other.frame?.shape
+      && this.frame?.fill === other.frame?.fill
+      && this.frame?.border === other.frame?.border
+      && this.frame?.borderWeight === other.frame?.borderWeight;
+
     return this.anchor.horizontal === other.anchor.horizontal && this.anchor.vertical === other.anchor.vertical
       && this.orientation.isAlmostEqual(other.orientation) && this.offset.isAlmostEqual(other.offset)
-      && this.textBlock.equals(other.textBlock);
+      && this.textBlock.equals(other.textBlock)
+      && framesMatch;
   }
 }
