+++ test-apps/display-test-app/src/frontend/TextDecoration.ts
@@ -3,10 +3,10 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 
-import { BaselineShift, ColorDef, FractionRun, GeometryStreamBuilder, IModelTileRpcInterface, LineBreakRun, TextAnnotation, TextAnnotationAnchor, TextBlock, TextBlockJustification, TextBlockMargins, TextRun, TextStyleSettingsProps } from "@itwin/core-common";
+import { BaselineShift, ColorDef, FractionRun, LineBreakRun, Placement2dProps, TextAnnotation, TextAnnotationAnchor, TextAnnotationFrameShape, TextAnnotationProps, TextBlock, TextBlockJustification, TextBlockMargins, TextFrameStyleProps, TextRun, TextStyleSettingsProps } from "@itwin/core-common";
 import { DecorateContext, Decorator, GraphicType, IModelApp, IModelConnection, readElementGraphics, RenderGraphicOwner, Tool } from "@itwin/core-frontend";
 import { DtaRpcInterface } from "../common/DtaRpcInterface";
-import { Guid, Id64, Id64String } from "@itwin/core-bentley";
+import { Id64, Id64String } from "@itwin/core-bentley";
 import { Point3d, YawPitchRollAngles } from "@itwin/core-geometry";
 
 // Ignoring the spelling of the keyins. They're case insensitive, so we check against lowercase.
@@ -24,11 +24,32 @@ class TextEditor implements Decorator {
   public rotation = 0;
   public offset = { x: 0, y: 0 };
   public anchor: TextAnnotationAnchor = { horizontal: "left", vertical: "top" };
+  public frame: TextFrameStyleProps = { borderWeight: 1, shape: "none" };
   public debugAnchorPointAndRange = false;
 
   // Properties applied to the entire document
   public get documentStyle(): Pick<TextStyleSettingsProps, "lineHeight" | "widthFactor" | "lineSpacingFactor"> {
     return this._textBlock.styleOverrides;
+  }
+
+  public get annotationProps(): TextAnnotationProps {
+    const annotation = TextAnnotation.fromJSON({
+      textBlock: this._textBlock.toJSON(),
+      // origin: this.origin,
+      anchor: this.anchor,
+      orientation: YawPitchRollAngles.createDegrees(this.rotation, 0, 0).toJSON(),
+      offset: this.offset,
+      frame: this.frame,
+    });
+
+    return annotation.toJSON();
+  }
+
+  public get placementProps(): Placement2dProps {
+    return {
+      origin: this.origin,
+      angle: 0,
+    }
   }
 
   // Properties to be applied to the next run
@@ -61,6 +82,7 @@ class TextEditor implements Decorator {
     this.debugAnchorPointAndRange = false;
     this.runStyle = { fontName: "Arial" };
     this.baselineShift = "none";
+    this.frame = { borderWeight: 1, shape: "none" };
   }
 
   public appendText(content: string): void {
@@ -109,6 +131,17 @@ class TextEditor implements Decorator {
     };
   }
 
+  public setFrame(frame: Partial<TextFrameStyleProps>) {
+    this.frame = { ...this.frame, ...frame };
+  }
+
+  /**
+   * Draws the graphics for the decoration. Text annotation graphics require a call to the backend to generate the geometry.
+   * In this case, we're using the `TextAnnotationGeometry` RPC endpoint that calls [[IModelDb.generateElementGraphics]]
+   * with the values from [[appendTextAnnotationGeometry]].
+   * These graphics can be added to the [[RenderSystem]] via [[readElementGraphics]] and [[RenderSystem.createGraphicOwner]]
+   * or via an [[ElementGeometryGraphicsProvider]]. In this case, we're using the former.
+   */
   public async update(): Promise<void> {
     if (!this._iModel) {
       throw new Error("Invoke `dta text init` first");
@@ -118,33 +151,15 @@ class TextEditor implements Decorator {
       return;
     }
 
-    const annotation = TextAnnotation.fromJSON({
-      textBlock: this._textBlock.toJSON(),
-      // origin: this.origin,
-      anchor: this.anchor,
-      orientation: YawPitchRollAngles.createDegrees(this.rotation, 0, 0).toJSON(),
-      offset: this.offset,
-    });
-
     const rpcProps = this._iModel.getRpcProps();
-    const geom = await DtaRpcInterface.getClient().produceTextAnnotationGeometry(rpcProps, annotation.toJSON(), this.debugAnchorPointAndRange);
-    const builder = new GeometryStreamBuilder();
-    builder.appendTextBlock(geom);
 
-    const gfx = await IModelTileRpcInterface.getClient().requestElementGraphics(rpcProps, {
-      id: Guid.createValue(),
-      toleranceLog10: -5,
-      type: "2d",
-      placement: {
-        origin: this.origin.toJSON(), // Point3d.createZero(),
-        angle: 0,
-      },
-      categoryId: this._categoryId,
-      geometry: {
-        format: "json",
-        data: builder.geometryStream,
-      },
-    });
+    const gfx = await DtaRpcInterface.getClient().generateTextAnnotationGeometry(
+      rpcProps,
+      this.annotationProps,
+      this._categoryId,
+      this.placementProps,
+      this.debugAnchorPointAndRange
+    );
 
     const graphic = undefined !== gfx ? await readElementGraphics(gfx, this._iModel, this._entityId, false) : undefined;
     this._graphic = graphic ? IModelApp.renderSystem.createGraphicOwner(graphic) : undefined;
@@ -335,6 +350,17 @@ export class TextDecorationTool extends Tool {
       }
       case "debug": {
         editor.debugAnchorPointAndRange = !editor.debugAnchorPointAndRange;
+        break;
+      }
+      case "frame": {
+        const key = inArgs[1];
+        const val = inArgs[2];
+        if (key === "fill") editor.setFrame({ fill: (val === "background" || val === "subcategory") ? val : val ? ColorDef.fromString(val).toJSON() : undefined });
+        else if (key === "border") editor.setFrame({ border: val ? ColorDef.fromString(val).toJSON() : undefined });
+        else if (key === "borderWeight") editor.setFrame({ borderWeight: Number(val) });
+        else if (key === "shape") editor.setFrame({ shape: val as TextAnnotationFrameShape });
+        else throw new Error("Expected shape, fill, border, borderWeight");
+
         break;
       }
